Added Logo
this was just to add a logo for now it will change down the rode to a better one. I just wanted it in place first and working then with refine later. There is still something weird with the top site filter I had to use negative margin and I want to get away from that.